+++ tasks.py
@@ -3,19 +3,19 @@ import os
 import sys
 import webbrowser
 
-from invoke import task, run
+from invoke import task
 
 docs_dir = 'docs'
 build_dir = os.path.join(docs_dir, '_build')
 
 @task
-def test(watch=False, last_failing=False):
+def test(ctx, watch=False, last_failing=False):
     """Run the tests.
 
     Note: --watch requires pytest-xdist to be installed.
     """
     import pytest
-    flake()
+    flake(ctx)
     args = []
     if watch:
         args.append('-f')
@@ -25,40 +25,40 @@ def test(watch=False, last_failing=False):
     sys.exit(retcode)
 
 @task
-def flake():
+def flake(ctx):
     """Run flake8 on codebase."""
-    run('flake8 .', echo=True)
+    ctx.run('flake8 .', echo=True)
 
 @task
-def clean():
-    run("rm -rf build")
-    run("rm -rf dist")
-    run("rm -rf flask_marshmallow.egg-info")
-    clean_docs()
+def clean(ctx):
+    ctx.run("rm -rf build")
+    ctx.run("rm -rf dist")
+    ctx.run("rm -rf flask_marshmallow.egg-info")
+    clean_docs(ctx)
     print("Cleaned up.")
 
 @task
-def clean_docs():
-    run("rm -rf %s" % build_dir)
+def clean_docs(ctx):
+    ctx.run("rm -rf %s" % build_dir)
 
 @task
-def browse_docs():
+def browse_docs(ctx):
     path = os.path.join(build_dir, 'index.html')
     webbrowser.open_new_tab(path)
 
 @task
-def docs(clean=False, browse=False, watch=False):
+def docs(ctx, clean=False, browse=False, watch=False):
     """Build the docs."""
     if clean:
-        clean_docs()
-    run("sphinx-build %s %s" % (docs_dir, build_dir))
+        clean_docs(ctx)
+    ctx.run('sphinx-build %s %s' % (docs_dir, build_dir))
     if browse:
-        browse_docs()
+        browse_docs(ctx)
     if watch:
-        watch_docs()
+        watch_docs(ctx)
 
 @task
-def watch_docs():
+def watch_docs(ctx):
     """Run build the docs when a file changes."""
     try:
         import sphinx_autobuild  # noqa
@@ -68,27 +68,27 @@ def watch_docs():
         print('    pip install sphinx-autobuild')
         sys.exit(1)
     docs()
-    run('sphinx-autobuild {} {}'.format(docs_dir, build_dir), pty=True)
+    ctx.run('sphinx-autobuild {} {}'.format(docs_dir, build_dir), pty=True)
 
 
 @task
-def readme(browse=False):
-    run('rst2html.py README.rst > README.html')
+def readme(ctx, browse=False):
+    ctx.run('rst2html.py README.rst > README.html')
     if browse:
         webbrowser.open_new_tab('README.html')
 
 @task
-def publish(test=False):
+def publish(ctx, test=False):
     """Publish to the cheeseshop."""
-    clean()
+    clean(ctx)
     try:
         __import__('wheel')
     except ImportError:
         print('wheel required. Run `pip install wheel`.')
         sys.exit(1)
     if test:
-        run('python setup.py register -r test sdist bdist_wheel', echo=True)
-        run('twine upload dist/* -r test', echo=True)
+        ctx.run('python setup.py register -r test sdist bdist_wheel', echo=True)
+        ctx.run('twine upload dist/* -r test', echo=True)
     else:
-        run('python setup.py register sdist bdist_wheel', echo=True)
-        run('twine upload dist/*', echo=True)
+        ctx.run('python setup.py register sdist bdist_wheel', echo=True)
+        ctx.run('twine upload dist/*', echo=True)